Saving changes to push the other branch to repo
